admin_panel/requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.2
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.10.0
django-ckeditor==6.5.1
django-grappelli==3.0.6
django-import-export==3.2.0
django-js-asset==2.0.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
Faker==18.10.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.5.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
tablib==3.4.0
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
